backend/requirements.txt
Django==2.2.16
django-filter
django-debug-toolbar==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
mixer==7.1.2
Pillow==9.5.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1